backend/requirements.txt
arrow==1.2.3
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.1.2
numpy==1.24.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.0
pylibmc==1.6.3
python-dateutil==2.8.2
python-memcached==1.59
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
transliterate==1.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
